Split makeString() and friends out into their own header https://bugs.webkit.org/show_bug.cgi?id=275795
Reviewed by NOBODY (OOPS!).

Moved `makeString`, `tryMakeString`, `makeAtomString`, `tryMakeAtomString`
and `makeStringByInserting` into a new header, wtf/text/MakeString.h.

This is a prerequisite to supporting `interleave()` in makeString,
as this allows makeString to depend on StringBuilder without a cycle.

Also removes remaining users of the StringOperators style of concatenation
as there were only a few users remaining.

- Source list elided -